+++ utils/js/index.ts
@@ -4,6 +4,7 @@
  *
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
+export * from './ClipboardUtils';
 export * from './DownloadUtils';
 export * from './HtmlUtils';
 export * from './LangUtils';